feat(backend): add the ingestion pipeline, job tracking and price-file import
- normalize: strict period, date, decimal and identifier parsing that refuses
  anything ambiguous instead of guessing
- quality: documented rules; structural problems reject (kept as issues with
  the raw record), plausibility problems flag, values are never corrected
- persistence: revisions instead of overwrites, gzip captures with SHA-256,
  issues linked to the rows they concern, series/instrument summaries
- jobs: statuses derived from item outcomes (completed, with warnings,
  partially failed, failed, cancelled); abandoned runs closed honestly
- economic pipeline: one transaction per series, circuit breaker after
  repeated outages, Ctrl+C recorded as cancelled
- prices: licensed CSV import with a validated manifest (licence,
  attribution, instrument identifiers, currency and adjustment checks)
- CLI: python -m app.ingestion catalog|run|import-prices|jobs|job with
  meaningful exit codes
- decimal_fits ignores insignificant trailing zeros without expanding
  huge exponents; bytes_received widened to BIGINT

+++ backend/app/models/ingestion.py
@@ -10,7 +10,7 @@ import uuid
 from datetime import date, datetime
 from typing import Any
 
-from sqlalchemy import JSON, ForeignKey, Integer, LargeBinary, String, Text, Uuid
+from sqlalchemy import JSON, BigInteger, ForeignKey, Integer, LargeBinary, String, Text, Uuid
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base, enum_column, utcnow
@@ -66,7 +66,7 @@ class IngestionJob(Base):
     warning_count: Mapped[int] = mapped_column(Integer, default=0)
     error_count: Mapped[int] = mapped_column(Integer, default=0)
     request_count: Mapped[int] = mapped_column(Integer, default=0)
-    bytes_received: Mapped[int] = mapped_column(Integer, default=0)
+    bytes_received: Mapped[int] = mapped_column(BigInteger, default=0)
     # A short, safe explanation of failures (no secrets, no stack traces).
     error_summary: Mapped[str | None] = mapped_column(Text)
